feat(transactions,budgets): support decimal amounts and calculator ops - Change Transaction.amount and Budget.totalAmount to double - Update JSON (g.dart) parsing to use toDouble() - Accept double in Constants.formatThousandFromInt - Parse doubles in HomeScreen and TransactionController - Add calculator logic (+, -, x, /, .), C, backspace, equals - Wrap create screen in FormBuilder with validate+submit - Regenerate freezed and JSON model files

+++ lib/core/constants/constants.dart
@@ -3,7 +3,7 @@ import 'package:intl/intl.dart';
 
 class Constants {
   static final DateFormat dateFormat = DateFormat('dd MMM y');
-  static String formatThousandFromInt(int number) {
+  static String formatThousandFromInt(double number) {
     return NumberFormat.decimalPattern('vi_VN').format(number);
   }
 }
